apps/api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.10.0
tiktoken==0.5.2
python-multipart==0.0.6
httpx==0.26.0
beautifulsoup4==4.12.3
requests==2.31.0
